+++ .env
@@ -1,4 +1,3 @@
 REACT_APP_CLIENT_ID="266624760782258186"
 REACT_APP_BASE_WEB_URL="https://skyra.netlify.com"
 REACT_APP_BASE_API_URL="https://api.skyra.pw"
-REACT_APP_BASE_CDN_URL="https://cdn.skyra.pw"
