+++ .env.development
@@ -1,4 +1,4 @@
 REACT_APP_CLIENT_ID="365184854914236416"
 REACT_APP_BASE_WEB_URL="http://localhost:3000"
 REACT_APP_BASE_API_URL="http://localhost:1234"
-REACT_APP_BASE_CDN_URL="https://cdn.skyra.pw"
+REACT_EDITOR=code
